code cleanup and improvements
refactor: code cleanup and improvements

+++ docker/i2b2-oracle/create_oracle_image.sh
@@ -5,25 +5,29 @@
 # Script Name: create_oracle_image.sh
 # Description: Creates, configures, and pushes a Docker image containing 
 #              an Oracle database pre-loaded with i2b2 demo data.
-# Usage:       sh create_oracle_image.sh <I2B2_DATA_ORACLE_TAG>
+# Usage:       bash create_oracle_image.sh <I2B2_DATA_ORACLE_TAG>
 # Expected Output: A committed Docker image for Oracle with i2b2 data loaded.
 # ==============================================================================
 
 # Enforce strict error handling:
 # -e: Exit immediately if a command returns a non-zero status.
-# -u: Treat unset variables as an error.
-set -eu
+set -e
 
-# Validate input arguments
-if [ -z "${1:-}" ]; then
-    echo "Error: Missing required argument I2B2_DATA_ORACLE_TAG."
-    echo "Usage: $0 <I2B2_DATA_ORACLE_TAG>"
-    exit 1
+#local build or CI build
+if [ "$CI" = "true" ]; then
+    echo "Running in GitHub Actions.."
+else
+    echo "Running Locally.."
+    echo "This script requires sudo access to install Ant ."
+    sudo apt update && sudo apt install -y ant 
+    export docker_username="local"
+    export docker_reponame="local"
 fi
 
-I2B2_DATA_ORACLE_TAG="$1"
+I2B2_DATA_ORACLE_TAG="${1:-local}"
 I2B2_CORE_SERVER_HOST="i2b2-core-server"
 I2B2_CORE_SERVER_PORT="8080"
+I2B2_DATA_PATH=$(pwd)/../..
 
 # Secure credentials using defaults that can be overridden by environment variables
 ORACLE_PWD="${ORACLE_PWD:-MyStrongPass123}"
@@ -37,7 +41,7 @@ docker run -d \
   --name oracle23 \
   -p 1521:1521 \
   -e ORACLE_PWD="$ORACLE_PWD" \
-  -v "/home/runner/work/i2b2-data/i2b2-data/:/i2b2" \
+  -v "$I2B2_DATA_PATH:/i2b2" \
   --network i2b2-net \
   container-registry.oracle.com/database/free:latest
 
@@ -53,16 +57,15 @@ docker exec -e ORACLE_PWD="$ORACLE_PWD" oracle23 bash -c 'sqlplus -s "sys/${ORAC
 
 docker_network_gateway_ip=$(docker network inspect i2b2-net -f '{{range .IPAM.Config}}{{.Gateway}}{{end}}')
 
-root="/home/runner/work/i2b2-data/i2b2-data/"
-cd "$root"
+cd "$I2B2_DATA_PATH"
 
 echo "Loading CRC Data..."
 CELL=i2b2demodata
-cd "$root/edu.harvard.i2b2.data/Release_1-8/NewInstall/Crcdata/"
+cd "$I2B2_DATA_PATH/edu.harvard.i2b2.data/Release_1-8/NewInstall/Crcdata/"
 sed -e "s|localhost|$docker_network_gateway_ip|g" \
     -e "s|PWD|$DEMO_PASS|g" \
     -e "s|USER_NAME|$CELL|g" \
-    "$root/docker/i2b2-oracle/db.properties" > db.properties
+    "$I2B2_DATA_PATH/docker/i2b2-oracle/db.properties" > db.properties
 cat db.properties
 
 ant -f data_build.xml create_crcdata_tables_release_1-8
@@ -71,32 +74,32 @@ ant -f data_build.xml db_demodata_load_data
 
 echo "Loading HIVE Data..."
 CELL=i2b2hive
-cd "$root/edu.harvard.i2b2.data/Release_1-8/NewInstall/Hivedata"
+cd "$I2B2_DATA_PATH/edu.harvard.i2b2.data/Release_1-8/NewInstall/Hivedata"
 sed -e "s|localhost|$docker_network_gateway_ip|g" \
     -e "s|PWD|$DEMO_PASS|g" \
     -e "s|USER_NAME|$CELL|g" \
-    "$root/docker/i2b2-oracle/db.properties" > db.properties
+    "$I2B2_DATA_PATH/docker/i2b2-oracle/db.properties" > db.properties
 ant -f data_build.xml create_hivedata_tables_release_1-8
 ant -f data_build.xml db_hivedata_load_data
 
 echo "Loading IM Data..."
 CELL=i2b2imdata
-cd "$root/edu.harvard.i2b2.data/Release_1-8/NewInstall/Imdata"
+cd "$I2B2_DATA_PATH/edu.harvard.i2b2.data/Release_1-8/NewInstall/Imdata"
 sed -e "s|localhost|$docker_network_gateway_ip|g" \
     -e "s|PWD|$DEMO_PASS|g" \
     -e "s|USER_NAME|$CELL|g" \
-    "$root/docker/i2b2-oracle/db.properties" > db.properties
+    "$I2B2_DATA_PATH/docker/i2b2-oracle/db.properties" > db.properties
 cat db.properties
 ant -f data_build.xml create_imdata_tables_release_1-8
 ant -f data_build.xml db_imdata_load_data 
 
 echo "Loading Metadata..."
 CELL=i2b2metadata
-cd "$root/edu.harvard.i2b2.data/Release_1-8/NewInstall/Metadata"
+cd "$I2B2_DATA_PATH/edu.harvard.i2b2.data/Release_1-8/NewInstall/Metadata"
 sed -e "s|localhost|$docker_network_gateway_ip|g" \
     -e "s|PWD|$DEMO_PASS|g" \
     -e "s|USER_NAME|$CELL|g" \
-    "$root/docker/i2b2-oracle/db.properties" > db.properties
+    "$I2B2_DATA_PATH/docker/i2b2-oracle/db.properties" > db.properties
 cat db.properties
 ant -f data_build.xml create_metadata_tables_release_1-8
 ant -f data_build.xml db_metadata_load_data 
@@ -106,14 +109,14 @@ ant -f data_build.xml db_metadata_load_data
 
 echo "Loading PM Data..."
 CELL=i2b2pm
-cd "$root/edu.harvard.i2b2.data/Release_1-8/NewInstall/Pmdata"
+cd "$I2B2_DATA_PATH/edu.harvard.i2b2.data/Release_1-8/NewInstall/Pmdata"
 sed -e "s|localhost|$docker_network_gateway_ip|g" \
     -e "s|PWD|$DEMO_PASS|g" \
     -e "s|USER_NAME|$CELL|g" \
-    "$root/docker/i2b2-oracle/db.properties" > db.properties
+    "$I2B2_DATA_PATH/docker/i2b2-oracle/db.properties" > db.properties
 
 echo "Replacing host:port in Pmdata/scripts/demo/pm_access_insert_data.sql.."
-sed -i "s|localhost:9090|$I2B2_CORE_SERVER_HOST:$I2B2_CORE_SERVER_PORT|g" "$root/edu.harvard.i2b2.data/Release_1-8/NewInstall/Pmdata/scripts/demo/pm_access_insert_data.sql"
+sed -i "s|localhost:9090|$I2B2_CORE_SERVER_HOST:$I2B2_CORE_SERVER_PORT|g" "$I2B2_DATA_PATH/edu.harvard.i2b2.data/Release_1-8/NewInstall/Pmdata/scripts/demo/pm_access_insert_data.sql"
 
 ant -f data_build.xml create_pmdata_tables_release_1-8
 ant -f data_build.xml create_triggers_release_1-8
@@ -121,20 +124,19 @@ ant -f data_build.xml db_pmdata_load_data
 
 echo "Loading Workplace Data..."
 CELL=i2b2workdata
-cd "$root/edu.harvard.i2b2.data/Release_1-8/NewInstall/Workdata"
+cd "$I2B2_DATA_PATH/edu.harvard.i2b2.data/Release_1-8/NewInstall/Workdata"
 sed -e "s|localhost|$docker_network_gateway_ip|g" \
     -e "s|PWD|$DEMO_PASS|g" \
     -e "s|USER_NAME|$CELL|g" \
-    "$root/docker/i2b2-oracle/db.properties" > db.properties
+    "$I2B2_DATA_PATH/docker/i2b2-oracle/db.properties" > db.properties
 ant -f data_build.xml create_workdata_tables_release_1-8
 ant -f data_build.xml db_workdata_load_data
 
-cd "$root"
-
-echo "Cleaning up directories..." #Space Issue in Github CI Pipeline
-rm -rf .git
-rm -rf edu.harvard.i2b2.data
-df -h
+if [ "$CI" = "true" ]; then
+    echo "Cleaning up i2b2-data repo to resolve space issues..." #Space Issue in Github CI Pipeline
+    rm -rf .git
+    rm -rf "$I2B2_DATA_PATH/edu.harvard.i2b2.data/"
+fi
 
 echo "Committing Docker image..."
 docker commit oracle23 "${docker_username}/${docker_reponame}:i2b2-data-oracle_${I2B2_DATA_ORACLE_TAG}"
